pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-vault"
version = "0.4.0"
description = "Sync Claude and ChatGPT conversations to Obsidian markdown files"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Mario Padilla"}
]
license = {text = "AGPL-3.0-or-later OR Commercial"}
keywords = ["claude", "chatgpt", "obsidian", "markdown", "sync", "ai", "conversations", "ollama"]
dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.0.0",
    "python-frontmatter>=1.0.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "types-requests>=2.31.0",
]

[project.scripts]
claude-vault = "claude_vault.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["claude_vault*"]

# Ruff configuration
[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint.isort]
known-first-party = ["claude_vault"]

# Mypy configuration
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Set to true for stricter checking
ignore_missing_imports = true
files = ["claude_vault/**/*.py"]

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--color=yes",
]